Guide user to enable this IM.
- Make ImeServiceUI primary implementation of ComposeUI, and AppUI (now AndroidUI) is decoration of ImeServiceUI (using kotlin's interface delegation)
- Temporary set settings destination to MainActivity (From system's settings App)
- Fix a crash if Backspace is pressed right after 1.
- Move AppUI out of ComposeUI and rename to AndroidUI (Maybe system IM settings is navigatable only on Android?)
- Move InputSystemConnection out of T9App and move to UI.
