+++ common/ui/src/androidMain/kotlin/com/github/kentvu/t9vietnamese/ui/AppUiPreview.kt
@@ -1,37 +1,35 @@
 package com.github.kentvu.t9vietnamese.ui
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.kentvu.t9vietnamese.KeypadEvent
-import com.github.kentvu.t9vietnamese.T9App
 import com.github.kentvu.t9vietnamese.UI
-import com.github.kentvu.t9vietnamese.lib.EnvironmentInteraction
 import com.github.kentvu.t9vietnamese.model.CandidateSelection
-import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
-import okio.FileSystem
-import okio.Source
 
 @Preview
 @Composable
 fun AppPreview() {
-    val ui = T9UI(
+    val ui = remember { AndroidUI(
         CoroutineScope(Dispatchers.Default),
         MutableStateFlow(
             UI.State(
                 initialized = true,
                 candidates = CandidateSelection.from(listOf("aa", "cc", "dd")),
-                confirmedText = "Test UI"
             ) { ev ->
                 when (ev) {
                     is KeypadEvent.CandidateSelect -> TODO()
                     KeypadEvent.CloseRequest -> TODO()
                     is KeypadEvent.KeyPress -> TODO()
                 }
-            }),
-        {}
-    )
+              }
+        ),
+        close = {},
+        launchSystemImSettings = {},
+        launchImePicker = {},
+    ) }
     ui.AppUi()
 }
